% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binarize.R
\name{binarize}
\alias{binarize}
\title{Binarize a correlation matrix.}
\usage{
binarize(c)
}
\arguments{
\item{c}{The correlation matrix.}
}
\value{
A binary adjacency matrix
}
\description{
Binarize a correlation matrix.
}
\examples{
binarize(graph)

}
\author{
Brandon Vaughan
}
