% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustcoef_signed_mult.R
\name{clustcoef_signed_mult}
\alias{clustcoef_signed_mult}
\title{Calculate signed clustering coeffecient for a multiple graphs}
\usage{
clustcoef_signed_mult(graphs, method = "Constantini")
}
\arguments{
\item{graphs}{A list of igraph objects or matrices.}

\item{method}{Either "Constantini" (the default) or "Zhang"}

\item{col.names}{A character vector of column (node) names.}

\item{row.names}{A character vector of row (subject) names.}
}
\value{
A matrix of clustering scores for each node and each subject
}
\description{
Calculate the signed clustering coeffecient (local transitivity) for a list of graphs using
either the Zhang & Horvath method or the Constantini & Perugini method.
}
\details{
Calculate the signed clustering coeffecient (local transitivity) for a single graph using
either the Zhang & Horvath (2005) method or the Constantini & Perugini (2014) method.
The local transitivity (or clustering coeffecient) of a node is the fraction of
edges a node forms with its neighbors out of the the number of edges it would
take to make complete triangles. This code was adapted from the code in the qgraph package
with modifications to make the output as in the matlab script
clustering_coef_w_sign in the Brain Connectivity Toolbox. The Onella method is excluded due to similarities with the Barrat (available as the local_clustcoef
function in this package) and Zhang methods.
}
\examples{
**##Not run**
clustcoef_signed_mult(graphs)
## End(**Not run**)

}
\references{
\href{https://sites.google.com/site/bctnet/measures/list}{Brain Connectivity Toolbox}

Costantini, G., & Perugini, M. (2014). Generalization of Clustering Coefficients to Signed Correlation Networks. PLoS ONE, 9(2), e88669. http://doi.org/10.1371/journal.pone.0088669

Zhang, B., & Horvath, S. (2005). A general framework for weighted gene co-expression network analysis. Statistical Applications in Genetics and Molecular Biology, 4(1).
}
\seealso{
\link[rsfcNet]{local_trans}
\link[igraph]{transitivity}
\link[rsfcNet]{clustcoef_signed}
}
\author{
Brandon Vaughan
}
