% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterative_louvain.R
\name{consensus.utility.func}
\alias{consensus.utility.func}
\title{Utility function for louvain signed algorithm}
\usage{
consensus.utility.func(merged_results)
}
\arguments{
\item{merged_results}{the merged results}

\item{iter}{the number of iterations. defaults to 100 but more may be desired.}
}
\description{
Get the consensus community memberships from multiple iterations of a louvain algorithm
}
\details{
Gets the consensus for modules
}
\examples{
**##Not run**
consensus.utility.func(merged_results)
 ## End(**Not run**)



}
\author{
Brandon Vaughan
}
\keyword{internal}
