% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/current_centr.R
\name{current_centr_mult}
\alias{current_centr_mult}
\title{Find current-flow closeness centrality for a list of graphs}
\usage{
current_centr_mult(graphs, col.names = NULL, row.names = NULL,
  parallel = FALSE, cores = NA)
}
\arguments{
\item{graphs}{A list of igraph objects or connectivity matrices}

\item{col.names}{The names of each column (node labels).}

\item{row.names}{The names of each row (subject).}

\item{parallel}{Should multiple cores be used? Defaults to FALSE. If TRUE, progress bar is not displayed. This is normal.}

\item{cores}{How many cores should be used? Defaults to recommended 1 less than number of CPU cores.}
}
\value{
A matrix contaning the centrality scores.
}
\description{
A variant of closeness centrality based on circuits
}
\details{
Current flow centrality was developed based on the properties of electrical circuits. It is a variant of closeness
centrality that imagines edges resistors and the nodes as relays between resistors. Weighted edge information
is taken into account as the absolute value. Non-zero edge are conceptually "conductors" while zero edges are
"resistors."

Current flow centrality is somewhat mathematically involved, but the final formula is

\eqn{\displaystyle{\sum_j A_{i,j} (v_{i}^{(s,t)} - v_{j}^{(s,t)}) = u_{i}^{(s,t)}}$}.

See the website in the references for a full explanation. The method involves calculating the laplacian representation
of a graph. This code is modified from the code in the centiserve package.
}
\examples{
current_centr_mult(graph)
}
\references{
Brandes U., Fleischer D. (2005) Centrality Measures Based on Current Flow. In: Diekert V., Durand B. (eds) STACS 2005. STACS 2005. Lecture Notes in Computer Science, vol 3404. Springer, Berlin, Heidelberg

https://www.sci.unich.it/~francesc/teaching/network/flowcentrality.html
}
\seealso{
\code{\link[rsfcNet]{closeness_centr}}
\code{\link[rsfcNet]{laplace_centr_mult}}
\code{\link[rsfcNet]{current_centr}}
}
\author{
Brandon Vaughan
}
