% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/degree_centr.R
\name{degree_mult}
\alias{degree_mult}
\title{Calculate degree centrality for multiple graphs.}
\usage{
degree_mult(graphs, col.names = NULL, row.names = NULL)
}
\arguments{
\item{graphs}{a list of igraph objects.}

\item{col.names}{The names of each column (node labels).}

\item{row.names}{The names of each row (subject).}
}
\value{
A matrix of the degree centralities of each node for each subject.
}
\description{
This function calculates the number of non-zero connections each node in a network has for a list of graphs.
}
\details{
Degree centrality is simply the sum of present connections of node i to each other jth node:

\eqn{D_i= \sum_{j \neq i}^N e_{ij}}

For the weighted version of degree, see the functions for calculating strength.
}
\examples{

# degree = degree_mult(graphs,row.names = subj_numbers,col.names = node_labels)

}
\references{
Fornito, A., Zalesky, A., & Bullmore, E. (2016). Node Degree and Strength. Chapter 4. Fundamentals of Brain Network Analysis, 115-136. doi:10.1016/B978-0-12-407908-3.00004-2
}
\seealso{
\code{\link[rsfcNet]{degree_centr}}
\code{\link[rsfcNet]{strength_multiple}}
}
\author{
Brandon Vaughan
}
