% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_fit.R
\name{dist_fit}
\alias{dist_fit}
\title{Fit a vector of variables to the best distribution}
\usage{
dist_fit(variable, only.best = TRUE)
}
\arguments{
\item{variable}{A vector of values describing a graph theory metric of interest.}

\item{only.best}{Return only the parameters for the best fitting distribution (default is TRUE) or all distributions (FALSE).}
}
\description{
Automatically detect which distributions are appropriate to the data using Maximum Likelihood Estimation, then find the log likelihood and return
the parameters of the best fitting distribution. For example, fit the degree distribution to see if a power law distribution best describes the data.
}
\details{
First the function detects whether or not any negative values are present. If so, distributions that require all
positive values are excluded. Next, the data are fit to the appropriate distributions and the log likelihood is used
to select the best fitting model. This function was primarily designed for assessing whether or not the degree
distribution or strength distribution of a graph follows the power law, but can also be used to fit distributions
to other metrics (ie, eigenvector centrality, clustering coefficient, etc).
}
\examples{
**## Not run:**
k = degree(graph)
distr = dist_fit(k)

s = strength(graph)
distr = dist(s)
## End(**Not run**)


}
\seealso{
\code{\link[rsfcNet]{strength_distribution}}
}
