% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eigen_centr.R
\name{eigen_centr_mult}
\alias{eigen_centr_mult}
\title{Calculate eigenvector centrality for multiple graphs.}
\usage{
eigen_centr_mult(graphs, col.names = NULL, row.names = NULL)
}
\arguments{
\item{graphs}{a list of igraph objects.}

\item{col.names}{The names of each column (node labels).}

\item{row.names}{The names of each row (subject).}
}
\value{
A matrix of the eigenvector centralities of each node for each subject.
}
\description{
This function calculates the eigenvector centrality for multiple graphs. This function
is essentially a port of the matlab script from the Brain Connectivity Toolbox (Rubinov & Sporns, 2010).
}
\details{
The eigenvector centrality is the eigenvector for each node corresponding to the largest eigenvalue of the matrix.

Eigenvector centrality differs from degree or strength.
A node with many connections does not necessarily have a high eigenvector
centrality. For example, a node may have many very weak connections that yield a large value for strength/degree.
Likewise, a node with high eigenvector centrality may have few connections but be well connected to a small number
of important nodes.
}
\examples{

\donttest{
eigencent = eigen_centr_mult(graphs,row.names = subj_numbers,col.names = node_labels)
}

}
\references{
\href{https://sites.google.com/site/bctnet/measures/list}{Brain Connectivity Toolbox}

Lohmann, G., Margulies, D. S., Horstmann, A., Pleger, B., Lepsien, J., Goldhahn, D., … Turner, R. (2010). Eigenvector Centrality Mapping for Analyzing Connectivity Patterns in fMRI Data of the Human Brain. PLoS ONE, 5(4), e10232. doi:10.1371/journal.pone.0010232

Fornito, A., Zalesky, A., & Bullmore, E. (2016). Centrality and Hubs. Chapter 5. Fundamentals of Brain Network Analysis, 137-161. doi:10.1016/b978-0-12-407908-3.00005-4

Rubinov, M. , Sporns, O. (2010) Complex network measures of brain connectivity: Uses and interpretations. NeuroImage 52:1059-69.
}
\seealso{
\code{\link[rsfcNet]{eigen_centr}}
\code{\link[rsfcNet]{leverage_centr_mult}}
}
\author{
Brandon Vaughan
}
