% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_measures.R
\name{fiedler_value}
\alias{fiedler_value}
\title{Calculate the fiedler value for a graph}
\usage{
fiedler_value(graph)
}
\arguments{
\item{graph}{A network as an igraph object}
}
\value{
A numeric value of the fiedler value.
}
\description{
This function calculates the fiedler value for a graph.
}
\details{
The Fiedler value is the second smallest eigenvalue of the laplacian representation
of a graph. The closer the Fiedler value is to zero the more easily the graph can be split into
separate components unconnected to each other. The Fiedler value is also known as the algebraic
connectivity of a graph (Mohar, 1991). Hence the fiedler value can be used as a measure of a network's
robustness to becoming disconnected. See Daianu et al (2014) for an application to neuroimaging.
}
\examples{

energy = fiedler_value(Graph)

}
\references{
Mohar, B., The Laplacian Spectrum of Graphs, in Graph Theory, Combinatorics, and Applications, Vol. 2, Ed. Y. Alavi, G. Chartrand, O. R. Oellermann, A. J. Schwenk, Wiley, 1991, pp. 871–898.

Daianu, M., Jahanshad, N., Nir, T. M., Leonardo, C. D., Jack, C. R., Weiner, M. W., … Thompson, P. M. (2014). Algebraic connectivity of brain networks shows patterns of segregation leading to reduced network robustness in Alzheimer’s disease. Computational Diffusion MRI: MICCAI Workshop, Boston, MA, USA, September 2014.  MICCAI Workshop on Computation Diffusion MRI (Boston, MA), 55–64. http://doi.org/10.1007/978-3-319-11182-7_6
}
\seealso{
\code{\link[rsfcNet]{laplace_centr_mult}}
\code{\link[rsfcNet]{delta_energy_mult}}
\code{\link[rsfcNet]{fiedler_value_mult}}
}
\author{
Brandon Vaughan
}
