% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_average_graph.R
\name{get_average_graph}
\alias{get_average_graph}
\title{Average together a list of connectivity matrices}
\usage{
get_average_graph(cormats, n = NULL, which = "graph")
}
\arguments{
\item{cormats}{A list of correlation matrices}

\item{n}{The sample size.}

\item{which}{can be "graph" to get an igraph object, "matrix" for a correlation matrix, or "both" for a list of both. Defaults to graph.}
}
\value{
A matrix, igraph object, or list.
}
\description{
Average together a list of correlation matrices to analyze the typical network properties of a given group (ie, teens vs adults)
}
\examples{
\donttest{
cormats = get_cor_matrices(scrubbed_ts_list) #Get the correlation matrices first
graphs = get_average_graph(cormats)
}

}
\author{
Brandon Vaughan
}
