% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cor_matrices.R
\name{get_cor_matrices}
\alias{get_cor_matrices}
\title{Obtain correlation matrices for each subject.}
\usage{
get_cor_matrices(scrubbed, method = "partial", n = NULL,
  n.nodes = NULL)
}
\arguments{
\item{scrubbed}{The list of scrubbed time series matrices.}

\item{method}{One of "pearson", "partial", or "covar".}

\item{n}{The number of subjects in the list. Checks the global environment for "n" but may be assigned directly.}

\item{n.nodes}{The number of nodes. Checks the global environment for "n.nodes" but may be assigned directly.}
}
\value{
The connectivity matrices.
}
\description{
Get regularized estimates of the correlation matrices using a variety of algorithms.
}
\details{
For methods "pearson", "partial", "covar", the estimation is done using the James-Stein based
methods described in Schaefer and Strimmer (2005) and the optimal shrinkage level calculated
for each subject using the method of Opgen-Rhein and Strimmer (2007). Respectively these give the
regularized pearson correlation matrix, shrinkage estimated partial correlation matrix, and regularized
covariance matrix, This depends on the corpcor package.The covariance option shouldn't be used directly
for network analysis, but can be used with whatever you need it for. As the number of observations increases
relative to the number of variables these all converge onto the empirical pearson, partial, and covariance
matrices respectively.
}
\examples{
\donttest{
cormats = get_cor_matrices(scrubbed_ts_list)
}

}
\references{
Opgen-Rhein, R., and K. Strimmer (2007). Accurate ranking of differentially expressed genes by a distribution-free shrinkage approach. Statist. Appl. Genet. Mol. Biol. 6:9. doi:10.2202/1544-6115.1252

Schaefer, J., and K. Strimmer (2005). A shrinkage approach to large-scale covariance estimation and implications for functional genomics. Statist. Appl. Genet. Mol. Biol. 4:32. doi:10.2202/1544-6115.1175
}
\seealso{
\code{\link[rsfcNet]{threshold_matrix}}
\code{\link[rsfcNet]{binarize}}
}
\author{
Brandon Vaughan
}
