% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_file_list.R
\name{get_file_list}
\alias{get_file_list}
\title{Imports List of Time Series or Confound Files}
\usage{
get_file_list(path, full.names = "TRUE")
}
\arguments{
\item{path}{The location of the files.}

\item{full.names}{TRUE gets full path. FALSE retrieves only the names of the files. Defaults to TRUE.}
}
\value{
The list of file locations or names.
}
\description{
This function stores the file locations of each subject's file in a list.
}
\examples{
ts_file_locations = get_file_list(path="C:/Users/Yourname/Documents/brainstudy/ts/")

confound_file_locations = get_file_list(path="C:/Users/Yourname/Documents/brainstudy/confoundm/")

}
\author{
Brandon Vaughan
}
