% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_clustcoef.R
\name{global_clust}
\alias{global_clust}
\title{Calculate the global average clustering coefficient for a single graph}
\usage{
global_clust(graph, method = "Constantini")
}
\arguments{
\item{graph}{A graph as an igraph object or matrix..}

\item{method}{Either "Constantini" (the default) or "Zhang"}
}
\value{
The global clustering coefficient for a graph
}
\description{
Calculates the average clustering coefficient for a single graph using any of the variants
of weighted, signed, clustering coefficient.
}
\details{
One straightofrward measure of the global clustering coefficient was put forth by Watts and Strogatz
(1998) as simply the average clustering coefficient across all nodes in a network.
}
\references{
Watts, D. J.  and Strogatz, S. (1998). Collective dynamics of 'small-world' networks. Nature. 393 (6684): 440–442. doi:10.1038/30918
}
\seealso{
\link[rsfcNet]{local_trans}
\link[igraph]{transitivity}
\link[rsfcNet]{clustcoef_signed}
}
\author{
Brandon vaughan
}
