% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_clustcoef.R
\name{global_clust_mult}
\alias{global_clust_mult}
\title{Calculate the global average clustering coefficient for a list of graphs.}
\usage{
global_clust_mult(graphs, method = "Constantini")
}
\arguments{
\item{graphs}{A list of igraph objects or connectivity matrices.}

\item{method}{Either "Constantini" (the default) or "Zhang"}
}
\value{
A vector of global clustering coefficients for the graphs
}
\description{
Calculates the average clustering coefficient for a list of graphs using any of the variants
of weighted, signed, clustering coefficient.
}
\details{
One straightofrward measure of the global clustering coefficient was put forth by Watts and Strogatz
(1998) as simply the average clustering coefficient across all nodes in a network.
}
\seealso{
\link[rsfcNet]{global_trans}
\link[rsfcNet]{global_clust}
\link[rsfcNet]{clustcoef_signed_mult}
}
\author{
Brandon vaughan
}
