% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_trans.R
\name{global_trans}
\alias{global_trans}
\title{Calculate the global transitivity of a list of graphs}
\usage{
global_trans(graphs)
}
\arguments{
\item{graphs}{a list of igraph objects.}
}
\value{
A vector of global transitivity scores for each subject
}
\description{
This function is a convenience wrapper for the transitivity function in igraph. It is used to calculate the global clustering coefficient of a list of unweighted graphs or a thresholded weighted graph
}
\details{
Global transitivity is the ratio of complete triangles to all triplets (three nodes connected by at least two triplets, ie a "V" shape or a complete triangle) in the network. It is a variant of the global clustering coefficient (Wasserman & Faust, 1994).
}
\examples{
GTR = global_trans(graphs)

}
\references{
Wasserman, S., and Faust, K. (1994). Social Network Analysis: Methods and Applications. Cambridge: Cambridge University Press.
}
\author{
Brandon Vaughan
}
