% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_measures.R
\name{graph_energy_mult}
\alias{graph_energy_mult}
\title{Calculate the graph energy for a list of graphs}
\usage{
graph_energy_mult(graphs)
}
\arguments{
\item{graphs}{A list of networks as igraph objects or as matrices}
}
\value{
A vector of graph energy values for each graph.
}
\description{
This function calculates the energy each graph in a list.
}
\details{
Graph energy was originally
applied in  organic chemistry to quantify the stability of molecular orbitals
associated with pi-electrons (Li, Shi, & Gutman, 2012). The graph energy informs about the connectivity of the graph
as a whole. A graph energy of zero means the nodes are not connected at all. The graph energy
is calculated by summing the absolute values of the eigenvalues for the matrix. See Daianu et al
(2015) for an example of an application to neuroimaging.
}
\examples{

energy = graph_energy_mult(Graphs)

}
\references{
Daianu, M., Mezher, A., Jahanshad, N., Hibar, D. P., Nir, T. M., Jack, C. R., … Thompson, P. M. (2015). SPECTRAL GRAPH THEORY AND GRAPH ENERGY METRICS SHOW EVIDENCE FOR THE ALZHEIMER’S DISEASE DISCONNECTION SYNDROME IN APOE-4 RISK GENE CARRIERS. IEEE International Symposium on Biomedical Imaging, 2015, 458–461. http://doi.org/10.1109/ISBI.2015.7163910

Li, X.; Shi, Y.; Gutman, I. (2012), Graph Energy, New York: Springer, ISBN 978-1-4614-4219-6.
}
\seealso{
\code{\link[rsfcNet]{laplace_centr_mult}}
\code{\link[rsfcNet]{delta_energy_mult}}
\code{\link[igraph]{fiedler_value_mult}}
}
\author{
Brandon Vaughan
}
