% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{heatmap}
\alias{heatmap}
\title{Visualize a connectivity matrix}
\usage{
heatmap(graph, plot = TRUE, type = "static", save.html = FALSE,
  plot.title = "Network Heatmap", directory = paste0(getwd(), "/"),
  filename = "heatmap.html", color.scheme = viridis::viridis(500))
}
\arguments{
\item{graph}{an igraph object or matrix}

\item{type}{"static" (the default) creates a standard image, "html" creates an interactive image with plotly}

\item{save.html}{Save the HTML file (only for type="html")}

\item{plot.title}{A title for the image}

\item{directory}{Where to save the HTML for a plotly image. Deaults to current working directory.}

\item{filename}{What to call the html file}

\item{color.scheme}{A color scheme. Defaults to viridis from the viridis package, a colorblind friendly palette.}
}
\description{
Visualize a connectivity matrix
}
\details{
This creates a visualization of a connectivity matrix. Either a standard static image
can be created or an interactive image with plotly can be created. Color palettes can be supplied
that will work with either output type with packages such as viridis, colorRamps, grDevices, and
RColorBrewer.

A few color palettes I recommend besides the default:

colorRamps::blue2red(500) , which is particuarly good for visualizing a contrast
between positive and negative correlations

colorRamps::ygobb(500))

colorRamps::matlab.like(500)

colorRamps::matlab.like2(500)

viridis::magma(500)

viridis::plasma(500)

grDevices::grey.colors(500)
}
\author{
Brandon Vaughan

This creates a correlogram (heatmap) with plotly in html format.
}
