% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterative_louvain.R
\name{iterative_louvain_signed}
\alias{iterative_louvain_signed}
\title{Reiterative louvain community detection for signed networks}
\usage{
iterative_louvain_signed(graph, iter = 100, cores = 1)
}
\arguments{
\item{graph}{an igraph object}

\item{iter}{the number of iterations. defaults to 100 but more may be desired.}
}
\description{
Get the consensus community memberships from multiple iterations of a louvain algorithm
}
\details{
Gets the consensus for modules
}
\examples{
**##Not run**
iterative_louvain_signed(graph)
 ## End(**Not run**)



}
\author{
Brandon Vaughan
}
