% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LASSO.R
\name{LASSO}
\alias{LASSO}
\title{Variable selection using LASSO (L1 penalty) regression.}
\usage{
LASSO(x, y, var.names = NULL, family = "gaussian", output = "html",
  return.glmnet = FALSE)
}
\arguments{
\item{x}{A matrix of predictor (independent) variables.}

\item{y}{A vector of response (dependent) variables. For gaussian models this will be a numeric vector. For poisson it will be a numeric vector of 0+ integers (count data). For binomial or multinomial regression this will be a vector containing factor labels. For mgaussian (multiple response linear regression; see ) y is a matrix of numeric response variables.}

\item{var.names}{The names of each node.}

\item{family}{Options include "gaussian", "poisson", "binomial" (logistic), "multinomial" (multiple outcome logistic), and "mgaussian"}

\item{output}{Options include "html" (the default), "markdown", "latex", or return to console.}

\item{return.glmnet}{If set to TRUE returns a copy of the cross-validated glmnet model to the global environment. Defaults to false. Set to TRUE if you want to test it on a holdout sample (only recommended if n is large).}
}
\value{
A table of variables that survived the variable selection.
}
\description{
This function is a convenience wrapper for glmnet's lasso capability. It will automatically select the best lambda value (strength of the L1 penalty) via cross-validation and return a table of variables that are good predictors of the response variable(s).
}
\examples{
lasso.results = LASSO(predictor_matrix, smoker_status, family="binomial")

}
\references{
Jerome Friedman, Trevor Hastie, Robert Tibshirani (2010). Regularization Paths for Generalized Linear Models via Coordinate Descent. Journal of Statistical Software, 33(1), 1-22. URL http://www.jstatsoft.org/v33/i01/.

Tibshirani, Robert (1996). Regression Shrinkage and Selection via the lasso. Journal of the Royal Statistical Society. Series B (methodological). Wiley. 58 (1): 267–88. https://www.jstor.org/stable/2346178
}
\author{
Brandon Vaughan
}
