% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mats_as_graphs.R
\name{mats_as_graphs}
\alias{mats_as_graphs}
\title{Convert list of connectivity matrices to igraph objects}
\usage{
mats_as_graphs(cormats)
}
\arguments{
\item{cormats}{a list of matrices}
}
\value{
A list of igraph objects
}
\description{
This function converts a list of adjacency or correlation matrices to igraph objects.
}
\examples{
graphs = mats_to_graphs(cormats)

}
\author{
Brandon Vaughan
}
