% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pagerank_centr.R
\name{pagerank_centr}
\alias{pagerank_centr}
\title{Calculate page_rank centrality for all nodes for multiple subjects.}
\usage{
pagerank_centr(graphs, damp = 0.85, prior.probs = "NULL",
  col.names = NULL, row.names = NULL)
}
\arguments{
\item{graphs}{a list of igraph objects.}

\item{damp}{= damping factor determining the probability that a random walk will continue from a given node. Defaults to .85}

\item{prior.probs}{= an optional list of vectors containing weights for the prior probability of each node being visited.}

\item{col.names}{The names of each column (node labels).}

\item{row.names}{The names of each row (subject).}
}
\value{
A matrix of the page_rank centralities of each node for each subject.
}
\description{
This function is a convenience wrapper for igraph's page_rank function.
}
\details{
The 'pagerank' centrality type results from a random walk of the network. At each node in the graph,
the next node is jumped to with probability p from the neighbors of node i.
Otherwise, or when a node has no successors, the next node is chosen from all nodes.
The centrality score is the average time spent at each node during the random walk. Pagerank assumes
equal probability for each node, but prior probabilities can be supplied.
}
\examples{
page_rank = pagerank_centr(graphs)

}
\references{
Brin, S., Page, L. (1998). The Anatomy of a Large-Scale Hypertextual Web Search Engine. Proceedings of the 7th World-Wide Web Conference, Brisbane, Australia.

Fornito, A., Zalesky, A., & Bullmore, E. (2016). Centrality and Hubs. Chapter 5. Fundamentals of Brain Network Analysis, 137-161. doi:10.1016/b978-0-12-407908-3.00005-4
}
\seealso{
\code{\link[rsfcNet]{eigen_centr}}
}
\author{
Brandon Vaughan
}
