% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_files.R
\name{save_files}
\alias{save_files}
\title{Save a list of numeric matrices}
\usage{
save_files(files, names, path = getwd())
}
\arguments{
\item{files}{a list of matrices with numeric data.}

\item{names}{a vector of names for files. Required.}

\item{path}{a path to save files in. Required, but if left blank saves to current working directory.}
}
\value{
returns nothing, but saves files to a directory.
}
\description{
Save a list of numeric matrices to the .csv file format.
}
\examples{
**## Not run:**
save_files(cormats, subj.names, path="C:/Users/YourName/Documents/R/")
## End(**Not run**)
}
\author{
Brandon Vaughan
}
