% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrub_matrix.R
\name{scrub_matrix}
\alias{scrub_matrix}
\title{Used in scrub_time_series function.}
\usage{
scrub_matrix(i)
}
\arguments{
\item{i}{The time series matrix}
}
\value{
The scrubbed time series matrix.
}
\description{
This function is used to scrub the ith matrix in a list. For use in scrub_time_series function.
}
\examples{
See scrub_time_series.
}
\author{
Brandon Vaughan
}
