% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shrinkage_ttest.R
\name{shrinkage_ttest}
\alias{shrinkage_ttest}
\title{Use a shrinkage estimated t-test over all pairwise comparisons.}
\usage{
shrinkage_ttest(variables, demographics, ngrp1 = NULL, ngrp2 = NULL,
  var.equal = FALSE, output = "full.console", bf.threshold = 5,
  alpha = 0.05, rscale = "medium")
}
\arguments{
\item{variables}{a data frame or matrix containing all dependent variables.}

\item{ngrp1}{Sample size of group 1.}

\item{ngrp2}{Sample size of group 2.}

\item{var.equal}{Should variances be assumed to be equal? Default is false, which maintains power when variances are equal anyway.}

\item{output}{Options equal print all evidence measures to the console (select this if you want to assign the results to a data frame), or "html.qvalue", "html.pvalue", and "html.bf" to print specific oututs to an html table. Replace html with latex for a latex table, with markdown for a markdown format table, or with console to print to the console (ie, "latex.bf", "markdown.pvalue", "console.qvalue")}

\item{bf.threshold}{The strength of evidence needed to threshold interesting results from uninteresting. Defaults to 5. 3 is minimum to be considered any evidence, 5 the minimum for substantial evidence, and 10 for very strong evidence.}

\item{alpha}{Significance level for p-values and q-values. Defaults to .05}

\item{rscale}{The prior distribution scale for expected effect size. Defaults to "wide". Set to "ultrawide" to set more stringent expectations of effect size, "medium" for a liberal threshold (perhaps not wise for big P scenarios). See documentation in BayesFactor package for more information.}

\item{demograhics}{a vector containing factor labels for each subject (the independent variable)}
}
\value{
A table of the significantly different nodes in a format of your choice with a definition of significance of your choice.
}
\description{
This function returns the pairwise comparisons for all nodes using the shrinkage t-test from the st package. The shrinkage t-statistic is a regularized estimate of the t-statistic for large P situations. It accounts for correlations between the independent variables such that an over-abundance of significant results are not driven by the correlation of variables. This approach was originally applied to microarray data (see Opgen-Rhein and Strimmer, 2007). This function extends the original by providing Bayes Factor tests, p-values, and q-values (FDR adjusted p-values). Bayes Factors calculated with BayesFactor package and q-values with fdrtools package. If n is less than half of P the Bayes Factors may be better to use, as q-values will be too conservative. Bayes Factors quantify evidence for or against a hypothesis and even a relatively low Bayes Factor threshold (such as BF > 3) is more conservative than a p<=.05 criterion (see Wetzels et al, 2011; Jeong and Boeck, 2017)
}
\examples{
shrinkage_ttest(eigen.centrality, demographics$gender, var.equal=FALSE, ngrp1=63, ngrp2=58, output="html.bf")
}
\references{
Opgen-Rhein, R., and K. Strimmer. 2007. Accurate ranking of differentially expressed genes by a distribution-free shrinkage approach. Statist. Appl. Genet. Mol. Biol. 6:9. http://dx.doi.org/10.2202/1544-6115.1252
Wetzels, R., Matzke, D., Lee, M. D., Rouder, J. N., Iverson, G. J., & Wagenmakers, E. (2011). Statistical Evidence in Experimental Psychology. Perspectives on Psychological Science, 6(3), 291-298. doi:10.1177/1745691611406923
Jeon, M., & Boeck, P. D. (2017). Decision qualities of Bayes factor and p value-based hypothesis testing. Psychological Methods, 22(2), 340-360. doi:10.1037/met0000140
}
\author{
Brandon Vaughan
}
