% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/softThreshold.R
\name{soft_thresh}
\alias{soft_thresh}
\title{Soft Thresholding Function for Signed Correlation Networks}
\usage{
soft_thresh(x, lambda = 8, zero.diag = T)
}
\arguments{
\item{x}{a correlation or partial correlation matrix}

\item{lambda}{the power to which the matrix will be raised. defaults to 8.
setting lambda to "auto" will automatically choose a value based on the operator norm of the transformed matrix.
A small value is added to lambda at each iteration, and the algorithm stops when the change in operator norm is <= than 0.01.}

\item{zero.diag}{if TRUE (the default) the diagonal of the returned matrix will be set to zero. otherwise, it will be set to 1.}
}
\value{
a matrix
}
\description{
This function implements the soft thresholding function from WGCNA. This is used to downweight negative
edges while at the same time shifting the matrix entries to be >= 0. This is accomplished by adding 1 to the matrix
entries, then dividing by two. Then, the resulting values are raised to a power, denoted as lambda.
Recommended values for lambda are 4 to 8 for partial correlation networks, and 8 to 12 for correlation networks. Larger
values result in stronger thresholding. The formula used is given below. \cr

\eqn{w_{ij}^{\text {signed }}=\left[\frac{\rho\left(i,j\right)+1}{2}\right]^{\lambda}}
}
