% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strength_distribution.R
\name{strength_distribution}
\alias{strength_distribution}
\title{Get the degree or strength distribution for a graph}
\usage{
strength_distribution(
  graph,
  cumulative = FALSE,
  signed = TRUE,
  which.sign = "star",
  count = FALSE
)
}
\arguments{
\item{graph}{A network as an igraph object or connectivity matrix.}

\item{cumulative}{Should the cumulative sum be returned or the density? Defaults to cumulative=FALSE}

\item{signed}{If TRUE (the default) uses the strength_signed function to calculate the distribution.}

\item{which.sign}{If "star" returns the weighted strength* metric, if "pos" returns the distribution of the
positive connections, and if "neg" returns the distribution of the negative connections.}

\item{count}{If the weights are binary can be set to count=TRUE to get the counts of each degree instead
of the probability densities.}
}
\description{
Get either the degree distribution (as a count or as a density) or the strength distribution for a graph.
}
\examples{
#' **## Not run:**
# If you want to examine the distributions of positive and negative connections
pos = strength_distribution(graph, which.sign="pos")
neg = strength_distribution(graph, which.sign="neg")

## End(**Not run**)

}
\seealso{
\code{\link[rsfcNet]{dist_fit}}
\code{\link[rsfcNet]{strength_signed}}
}
\author{
Brandon Vaughan
}
