% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strength_multiple.R
\name{strength_multiple}
\alias{strength_multiple}
\title{Calculate node connection strength for weighted undirected networks (multiple subjects).}
\usage{
strength_multiple(cormats, which = "all", Scale = "TRUE",
  col.names = .GlobalEnv$colnames, row.names = .GlobalEnv$rownames)
}
\arguments{
\item{cormats}{a list of connectivity matrices.}

\item{which}{which = "all" returns the strength star metric from Rubinov and Sporns (2011) along with the positive strengths and negative strengths for each node. which="star", "pos", and "neg" return only the weighted strength, positive strength, and negative strength respectively.}

\item{Scale}{Should Strength* be scaled to the range -1 to 1 (assuming correlations are the edge weights)? Defaults to Scale="TRUE"}

\item{col.names}{The names of each column (node labels)}

\item{row.names}{The names of each row (subject)}
}
\value{
A matrix of the statistic for the chosen strength measure, or a list of matrices if which="all"
}
\description{
This calculates the strength star metric proposed by Rubinov and Sporns (2011). Strength star is a weighted combination of the positive connections and negative connections in a correlation matrix. This function can also return positive and negative strength.
}
\examples{
strength = strength_multiple(cormats, which="star")
}
\references{
Rubinov, M., & Sporns, O. (2011). Weight-conserving characterization of complex functional brain networks. NeuroImage, 56(4), 2068-2079. doi:10.1016/j.neuroimage.2011.03.069
}
\author{
Brandon Vaughan
}
