% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strength_star.R
\name{strength_star}
\alias{strength_star}
\title{Calculate node connection strength for weighted undirected networks.}
\usage{
strength_star(matrix, scale = TRUE)
}
\arguments{
\item{matrix}{a single connectivity matrix .}

\item{scale}{Whether or not to scale the strength of each node into a 0-1 range (TRUE) or simply return the weighted sum of connections (FALSE).}
}
\value{
The scrubbed time series matrix.
}
\description{
This calculates the strength star metric proposed by Rubinov and Sporns (2011). Strength star is a weighted combination of the positive connections and negative connections in a correlation matrix.
}
\examples{
strength_star(sub001_cormat, scale=TRUE)
}
\references{
Rubinov, M., & Sporns, O. (2011). Weight-conserving characterization of complex functional brain networks. NeuroImage, 56(4), 2068-2079. doi:10.1016/j.neuroimage.2011.03.069
}
\author{
Brandon Vaughan
}
