% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threshold_and_binarize.R
\name{threshold_and_binarize}
\alias{threshold_and_binarize}
\title{Binarize a correlation matrix.}
\usage{
threshold_and_binarize(c, q = 0.9, as.igraph = FALSE, pos.only = TRUE)
}
\arguments{
\item{c}{The correlation matrix.}

\item{q}{The quantile to threshold. Defaults to .90, such that all below the 90th quantile are set to zero.}

\item{pos.only}{Should negative connections be removed before thresholding? Defaults to TRUE.}
}
\value{
A binary adjacency matrix
}
\description{
Threshold a correlation matrix and convert it to an un-weighted (binary) adjacency matrix.
}
\examples{
threshold_and_binarize(matrix, q=.90) # This should threshold such that only positive values are returned.
threshold_and_binarize(abs(matrix), q=.90) # Taking the absolute value of the matrix will allow negative correlations sufficiently strong to be included as an edge.
}
\author{
Brandon Vaughan
}
