% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threshold_matrix.R
\name{transform_matrix}
\alias{transform_matrix}
\title{Transform a correlation matrix into cohen's d or z-scores and back}
\usage{
transform_matrix(c, method = "cohens.d")
}
\arguments{
\item{c}{The network as a matrix}

\item{method}{one of "cohens.d", "undo.d"", "fishers.z" or "undo.d""}
}
\value{
The transformed correlation matrix.
}
\description{
This function is used to transform correlation matrices into other units such as Cohen's d or Fisher's z.
}
\examples{
transform_matrix(matrix, method="cohens.d") # This should threshold such that only positive values are returned.
}
\author{
Brandon Vaughan
}
