% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vitality.R
\name{vitality}
\alias{vitality}
\title{Find the closeness vitality for a single graph.}
\usage{
vitality(graph, method = "regular")
}
\arguments{
\item{graph}{A network as an igraph object or connectivity matrix.}

\item{method}{One of "regular" (the default) or "current"}
}
\value{
A numeric vector contaning the vitality scores for each node.
}
\description{
Closeness vitality of a node is the change in the sum of distances between all node pairs when that node is removed from the network.
}
\details{
Closeness vitality of a node is the change in the sum of distances between all node pairs when that node is removed from the network.
Either traditional closeness centrality can be used to define the sum of distances (Wiener Index) or the current closeness
centrality metric can be substituted with method="current". This function uses the absolute value of the weights, so if you require
only positive edges to be considered you must threshold them out of the network.

The formula is fairly straightforward:

First compute the Wiener Index for the entire graph, which is the sum of the reciprocals of the closeness
centrality for each node:


\eqn{W=\sum_{i=n}^N \frac{1}{ClC_n}}


Next, for each node, calculate the Wiener index again for a subgraph including all nodes but node_i. The vitality
of the node is simply the difference between the total Wiener index and the Wiener index for the subgraph.


\eqn{Vitality = W_{total} - W_{subgraph}}


The vitality of a node measures the increase in transport cost for the whole network
when a node is dropped. Negative values indicate a node can be excluded and decrease costs.
As the name suggests, the metric measures who in a network is "vital to the operation" and
who is an expense to the network (Brandes & Erlebach, 2005). Also see \code{\link[rsfcNet:laplace_centr_mult]{laplacian centrality}}
for another measure of node centrality based on the impact to the network after deletion.
}
\examples{

**## Not run:**
vitality(graph)
## End(**Not run**)

}
\references{
Brandes, U. & Erlebach, T. (2005). Network Analysis: Methodological Foundations, U.S. Government Printing Office.

Brandes U., Fleischer D. (2005) Centrality Measures Based on Current Flow. In: Diekert V., Durand B. (eds) STACS 2005. STACS 2005. Lecture Notes in Computer Science, vol 3404. Springer, Berlin, Heidelberg
}
\seealso{
\code{\link[rsfcNet]{closeness_centr}}
\code{\link[rsfcNet]{current_centr}}
\code{\link[rsfcNet]{laplace_centr}}
}
\author{
Brandon Vaughan
}
