% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_brainNet.R
\name{write_brainNet}
\alias{write_brainNet}
\title{Create Node and Edge Files for Brain Net Viewer}
\usage{
write_brainNet(MNIcoordinates, graph, node.colors = rep(1,
  .GlobalEnv$n.nodes), node.size, node.labels = seq(1:.GlobalEnv$n.nodes),
  directory = getwd(), name = Sys.Date())
}
\arguments{
\item{MNIcoordinates}{A matrix with columns consisting of the X,Y,Z MNI152 coordinates (must be in this order) for ROIs}

\item{graph}{A network in either matrix format or as an igraph object}

\item{node.colors}{A vector of labels, such as module membership. Defaults to all the same color.}

\item{node.size}{A vector of numbers describing the nodes (such as degree)}

\item{directory}{Directory for file output. Defaults to current working directory.}

\item{name}{A name for the output. Defaults to the current date.}
}
\description{
Create Node and Edge Files for Brain Net Viewer
}
\examples{
brainnet.matrix = threshold_matrix(avg_matrix) # Threshold a weighted correlation matrix first
brainnet.matrix = threshold_and_binarize(avg_matrix) # Or binarize
write.brainNet(MNIcoordinates=Node_Atlas[,1:3], graph = brainnet.matrix, node.colors=module$membership, node.size=LEV.AVG, directory="C:/Users/Abnormally-Distributed/Documents/R/", name="leverage_network")

}
\references{
Xia M, Wang J, He Y (2013) BrainNet Viewer: A Network Visualization Tool for Human Brain Connectomics. PLoS ONE 8: e68910.


https://www.nitrc.org/projects/bnv/
}
\author{
Brandon Vaughan
}
