% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_gephi.R
\name{write_gephi}
\alias{write_gephi}
\title{Create Gephi File from graph object}
\usage{
write_gephi(graph, node_attributes, directory = getwd(), name = Sys.Date(),
  node.names = .GlobalEnv$colnames)
}
\arguments{
\item{graph}{An igraph object}

\item{directory}{Character vector specifyig directory for file output. Defaults to current working directory.}

\item{name}{A name for the output. Defaults to the current date.}

\item{node.names}{A vector of node labels. By default searches global environemnt for a vector titled "colnames".}

\item{node.attributes}{A vector or data frame of different metrics describing the nodes (ie, modularity, degree).}
}
\description{
Create Gephi File from graph object
}
\details{
This function simplifies the use of the igraph to .gexf conversion in the rgexf package.
}
\examples{
attributes = cbind.data.frame(colMeans(EVC) , colMean(STR), colMeans(MODULE))
write_gephi(graph=group_avg_graph, attributes, directory = "C:/Users/Abnormally-Distributed/Documents/", name="Group_Average_Graph")

}
\references{
Bastian M., Heymann S., Jacomy M. (2009). Gephi: an open source software for exploring and manipulating networks. International AAAI Conference on Weblogs and Social Media.

https://gephi.org/
}
\author{
Brandon Vaughan
}
