% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/degree_centr.R
\name{degree_centr}
\alias{degree_centr}
\title{Calculate degree centrality for a single graph.}
\usage{
degree_centr(graph)
}
\arguments{
\item{graph}{A network as an igraph object or connectivity matrix.}
}
\value{
A vector of centrality scores for each node.
}
\description{
This function calculates the number of non-zero connections each node in a network has for a single graph.
}
\details{
Degree centrality is simply the sum of present connections of node i to each other jth node:

\eqn{D_i= \sum_{j \neq i}^N e_{ij}}

For the weighted version of degree, see the functions for calculating strength.
}
\examples{

# degree = degree_centr(graph)

}
\references{
Fornito, A., Zalesky, A., & Bullmore, E. (2016). Node Degree and Strength. Chapter 4. Fundamentals of Brain Network Analysis, 115-136. doi:10.1016/B978-0-12-407908-3.00004-2
}
\seealso{
\code{\link[rsfcNet]{degree_mult}}
\code{\link[rsfcNet]{strength_signed}}
\code{\link[igraph]{strength}}
}
\author{
Brandon Vaughan
}
