% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eigen_centr.R
\name{eigen_centr}
\alias{eigen_centr}
\title{Calculate eigenvector centrality for all nodes for multiple subjects.}
\usage{
eigen_centr(graphs, col.names = NULL, row.names = NULL)
}
\arguments{
\item{graphs}{a list of igraph objects.}

\item{col.names}{The names of each column (node labels).}

\item{row.names}{The names of each row (subject).}
}
\value{
A matrix of the eigenvector centralities of each node for each subject.
}
\description{
This function is a convenience wrapper for igraph's eigenvector centrality function and takes as an input a list of igraph objects. ECV is calculated on the absolute values of the correlation matrix per Lohmann et al, 2010. ECV is standardized subject-wise to 0-1 range for consistency.
}
\examples{
eigencent = eigen_centr(graphs,row.names = subj_numbers,col.names = node_labels)

}
\references{
Lohmann, G., Margulies, D. S., Horstmann, A., Pleger, B., Lepsien, J., Goldhahn, D., … Turner, R. (2010). Eigenvector Centrality Mapping for Analyzing Connectivity Patterns in fMRI Data of the Human Brain. PLoS ONE, 5(4), e10232. doi:10.1371/journal.pone.0010232

Fornito, A., Zalesky, A., & Bullmore, E. (2016). Centrality and Hubs. Chapter 5. Fundamentals of Brain Network Analysis, 137-161. doi:10.1016/b978-0-12-407908-3.00005-4
}
\author{
Brandon Vaughan
}
