% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eigen_centr.R
\name{eigen_centr}
\alias{eigen_centr}
\title{Calculate eigenvector centrality for a single graph.}
\usage{
eigen_centr(graph)
}
\arguments{
\item{graph}{A network as an igraph object or matrix.}
}
\value{
A vector of centrality scores for each node.
}
\description{
This function calculates the eigenvector centrality for a single graph. This function
is essentially a port of the matlab script from the Brain Connectivity Toolbox (Rubinov & Sporns, 2010).
}
\details{
The eigenvector centrality is the eigenvector for each node corresponding to the largest eigenvalue of the matrix.

Eigenvector centrality differs from degree or strength.
A node with many connections does not necessarily have a high eigenvector
centrality. For example, a node may have many very weak connections that yield a large value for strength/degree.
Likewise, a node with high eigenvector centrality may have few connections but be well connected to a small number
of important nodes.
}
\examples{

\donttest{
eigencent = eigen_centr(graph)
}

}
\references{
\href{https://sites.google.com/site/bctnet/measures/list}{Brain Connectivity Toolbox}

Lohmann, G., Margulies, D. S., Horstmann, A., Pleger, B., Lepsien, J., Goldhahn, D., … Turner, R. (2010). Eigenvector Centrality Mapping for Analyzing Connectivity Patterns in fMRI Data of the Human Brain. PLoS ONE, 5(4), e10232. doi:10.1371/journal.pone.0010232

Fornito, A., Zalesky, A., & Bullmore, E. (2016). Centrality and Hubs. Chapter 5. Fundamentals of Brain Network Analysis, 137-161. doi:10.1016/b978-0-12-407908-3.00005-4

Rubinov, M. , Sporns, O. (2010) Complex network measures of brain connectivity: Uses and interpretations. NeuroImage 52:1059-69.
}
\seealso{
\code{\link[rsfcNet]{leverage_centr}}
\code{\link[rsfcNet]{eigen_centr_mult}}
}
\author{
Brandon Vaughan
}
