% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cor_matrices.R
\name{get_cor_matrices}
\alias{get_cor_matrices}
\title{Obtain correlation matrices for each subject.}
\usage{
get_cor_matrices(scrubbed_ts_bulk, method = "partial", n = .GlobalEnv$n,
  n.nodes = .GlobalEnv$n.nodes)
}
\arguments{
\item{scrubbed_ts_bulk}{The list of scrubbed time series matrices.}

\item{n}{The number of subjects in the list. Checks the global environment for "n" but may be assigned directly.}

\item{n.nodes}{The number of nodes. Checks the global environment for "n.nodes" but may be assigned directly.}
}
\value{
The scrubbed time series matrices.
}
\description{
This function is a convenience wrapper for various functions in the corpcor package. The default is the partial correlation matrix (method="partial") but the covariance matrix (method="covariance"), partial covariance (method="partial.var"), and regular Pearson correlation matrix (method="pearson") are also available options.  The estimation is done using the method described in Schafer and Strimmer (2005) and the optimal shrinkage level calculated for each subject using the method of Opgen-Rhein and Strimmer (2007).
}
\examples{
cormats = get_cor_matrices(scrubbed_ts_list)

}
\references{
Opgen-Rhein, R., and K. Strimmer. 2007. Accurate ranking of differentially expressed genes by a distribution-free shrinkage approach. Statist. Appl. Genet. Mol. Biol. 6:9. doi:10.2202/1544-6115.1252


Schaefer, J., and K. Strimmer. 2005. A shrinkage approach to large-scale covariance estimation and implications for functional genomics. Statist. Appl. Genet. Mol. Biol. 4:32. doi:10.2202/1544-6115.1175
}
\author{
Brandon Vaughan
}
