% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_from_list.R
\name{import_from_list}
\alias{import_from_list}
\title{Imports Time Series or Confound Files from the list created with get_file_list.}
\usage{
import_from_list(file.list)
}
\arguments{
\item{file.list}{The list of files.}
}
\value{
The time series matrix or confound matrix for each subject.
}
\description{
This function loads the files into the global environment as a list of matrices. The function expects each file to be a .csv.
}
\examples{
ts = import_from_list(ts_file_locations)

confounds = import_from_list(confound_file_locations)

}
\author{
Brandon Vaughan
}
