% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_from_list.R
\name{import_from_list}
\alias{import_from_list}
\title{Imports Time Series or Confound Files from the list created with get_file_list.}
\usage{
import_from_list(file.list, header, sep = "", matrix = TRUE)
}
\arguments{
\item{file.list}{The list of files.}

\item{header}{Are the cells in the top row of the spreadsheet colum names? Must be TRUE or FALSE.}

\item{sep}{The separator used to separate entries. Defaults to "" since this is the output format from fsl.}

\item{matrix}{Should the data be imported directly to a matrix? Defaults to TRUE.}
}
\value{
The time series matrix or confound matrix for each subject.
}
\description{
This function loads the files into the global environment as a list of matrices. The function expects each file to be in a .csv or .txt format.
}
\examples{
ts = import_from_list(ts_file_locations, header=TRUE)

confounds = import_from_list(confound_file_locations, header=FALSE)

}
\author{
Brandon Vaughan
}
