% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modularity_consensus.R
\name{modularity_consensus}
\alias{modularity_consensus}
\title{Get the consensus community membership labels}
\usage{
modularity_consensus(modules)
}
\arguments{
\item{modules}{a matrix containing membership labels across each row for each node (represented by column).}
}
\description{
Get the consensus community memberships for multiple graphs or multiple iterations of an algorithm over a single graph.
}
\details{
Gets the consensus for  modules
}
\examples{
**##Not run**
modularity_consensus(module_list)
 ## End(**Not run**)

}
\author{
Brandon Vaughan
}
