% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_connectivity.R
\name{module_connectivity}
\alias{module_connectivity}
\title{Calculate several module/community based network statistics}
\usage{
module_connectivity(graph, modules, n.nodes = NULL,
  consensus.output = FALSE)
}
\arguments{
\item{graph}{a single network in matrix format or as an igraph object}

\item{modules}{a communities object from an igraph clustering function}

\item{n.nodes}{The The number of nodes in each network.}
}
\value{
A matrix containing several modularity/community structure based centrality measures.
}
\description{
This calculates the (within-module) strength star and diversity star metrics proposed by Rubinov and Sporns (2011) along with several other metrics.
}
\details{
This calculates the within module strength (weighted degree) z-score (using the weighted
signed "strength star" metric) and "diversity star" metric proposed by Rubinov and Sporns (2011).
This also returns the participation coefficient weighted by positive and negative connections in the
same manner as diversity. Also returned are the percentage of positive/negative weighted degree
a node has within-module out of its total weighted degree.
A node that is well-integrated into a community/module should arguably have
a high percentage of its total positive connections inside the module,
and a relatively low percentage of its negative connections within the module.


The formula used for diversity:

\eqn{h_i = -\frac{1}{log(m)}  \sum_{u=1}^{N_M} \Bigg( \left ( \frac{s_{iu}}{s_i} \right ) \cdot log  \left ( \frac{s_{iu}}{s_i} \right ) \Bigg)}

The formula used for participation:

\eqn{p_i = 1 - \sum_{s=1}^{N_M} \left ( \frac{s_{iu}}{s_i} \right )^2}


The weighting for diversity is calculated as follows (and participation the same formula):

\eqn{h_{i}^{*} = h_{i}^{+} - \Bigg( \frac{s_i^{-}}{s_i^{+}+s_i^{-}}\Bigg) h_{i}^{-}}
}
\examples{
\dontrun{
module_stats = module_connectivity(graph, module)
}

}
\references{
Power, J. D., Schlaggar, B. L., Lessov-Schlaggar, C. N., & Petersen, S. E. (2013). Evidence for hubs in human functional brain networks. Neuron, 79(4), 10.1016/j.neuron.2013.07.035

Rubinov, M., & Sporns, O. (2011). Weight-conserving characterization of complex functional brain networks. NeuroImage, 56(4), 2068-2079. doi:10.1016/j.neuroimage.2011.03.069
}
\author{
Brandon Vaughan
}
