% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_connectivity.R
\name{module_connectivity}
\alias{module_connectivity}
\title{Calculate several module/community based network statistics}
\usage{
module_connectivity(connectivity_matrix, modules, scale = FALSE,
  n.nodes = .GlobalEnv$n.nodes)
}
\arguments{
\item{connectivity_matrix}{a single connectivity matrix .}

\item{modules}{The list of communities objects from an igraph communities detection function.}

\item{scale}{Whether or not to scale the strength of each node into a 0-1 range (TRUE) or simply return the weighted sum of connections (FALSE).}

\item{n.nodes}{The The number of nodes in each network.}
}
\value{
A matrix containing several modularity/community structure based centrality measures.
}
\description{
This calculates the (within-module) strength star and diversity star metrics proposed by Rubinov and Sporns (2011) along with several other metrics.
}
\details{
This calculates the (within-module) strength star and diversity star metrics proposed by Rubinov and Sporns (2011). Also returns Within-Module Z-Score, Within-Module Pos/Neg Strength, Diversity-Star, Participation Coefficient, and percentage of total connections that exist within-module. Z-Scores and Participation Coefficient measures are based on the strength (star) of the nodes. In the event a binary network is fed, diversity will not be calculated. Participation will be the more appropriate measure. The strength in the output for a binary graph is just the degree.
}
\examples{
module_statistics = module_connectivity(cormat, module_labels)

}
\references{
Power, J. D., Schlaggar, B. L., Lessov-Schlaggar, C. N., & Petersen, S. E. (2013). Evidence for hubs in human functional brain networks. Neuron, 79(4), 10.1016/j.neuron.2013.07.035

Rubinov, M., & Sporns, O. (2011). Weight-conserving characterization of complex functional brain networks. NeuroImage, 56(4), 2068-2079. doi:10.1016/j.neuroimage.2011.03.069
}
\author{
Brandon Vaughan
}
