% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strength_signed.R
\name{strength_signed}
\alias{strength_signed}
\title{Calculate signed strength statistics on a single graph.}
\usage{
strength_signed(matrix, scale = TRUE)
}
\arguments{
\item{matrix}{A network in matrix format or as an igraph object.}

\item{scale}{Defaults to TRUE to return the scaled statistic. Can be set to FALSE.}
}
\value{
The weighted strength metric along with the positive and negative strength.
}
\description{
This calculates strength statistics on a single graph.
}
\details{
Strength star is the weighted combination of the positive connections and negative connections
in a network proposed by Rubinov and Sporns (2011). The statistic is calculated by the following formula:
\eqn{s_i^{*}=s_i^+-\left(\frac{s_i^-}{s_i^++s_i^-}\right)s_i^-}
When normalized to the -1 to 1 interval with scale=TRUE the positive and negative strength
are normalized with the following formula first then plugged into the above formula:
\eqn{s_i^\pm = \frac1{n-1}s_i^\pm}
}
\examples{
strength_star(sub001_cormat, scale=TRUE)
}
\references{
Rubinov, M., & Sporns, O. (2011). Weight-conserving characterization of complex functional brain networks. NeuroImage, 56(4), 2068-2079. doi:10.1016/j.neuroimage.2011.03.069
}
\author{
Brandon Vaughan
}
