% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threshold_matrix.R
\name{threshold_matrix}
\alias{threshold_matrix}
\title{Threshold a correlation matrix}
\usage{
threshold_matrix(c, q.thresh = 0.9, thresh)
}
\arguments{
\item{c}{The network either as connectivity matrix or as an igraph object.}

\item{q.thresh}{The quantile to threshold. Defaults to .90, such that all below the 90th quantile are set to zero.}

\item{thresh}{The number (taken to be an absolute value) below which to threshold (ie, correlation or cohen's d , etc). No default. Recommend .10 for correlation or .20 for cohen's d.}
}
\value{
The pruned correlation matrix.
}
\description{
This function is used to threshold a correlation matrix in order to prune weak connections.
}
\examples{
threshold_matrix(matrix, q=.90) # This should threshold such that only positive values are returned.
threshold_matrix(abs(matrix), q=.90) # Taking the absolute value of the matrix will allow negative correlations sufficiently strong to be included as an edge. This does not preserve signs.
threshold_matrix(abs(matrix), q=.90)*sign(matrix) This will restore the signs after thresholding if desired.

}
\author{
Brandon Vaughan
}
